requirements.txt
Authlib == 1.6.1
cs50 == 9.4.0
Flask == 3.1.1
Flask-session == 0.8.0
gunicorn == 23.0.0
google-api-core == 2.25.1
google-api-python-client == 2.181.0
google-auth == 2.40.3
google-auth-httplib2 == 0.2.0
google-auth-oauthlib == 1.2.2
Jinja2 == 3.1.6
oauthlib == 3.3.1
python-dotenv == 1.1.1
requests-oauthlib == 2.0.0
Werkzeug == 3.1.3